backend/requirements.txt
websockets==14.1
asyncio==3.4.3
google-cloud-logging